Fixed some bugs and added some features
Fixed not opening the application and added another option to search for "pluralsight.exe". It chooses the latest version of pluralsight and opens it using the same user path. It checks twice whether the username is the same as the path or not. Because for some reasons, it can be different, which will cause "pluralsight.exe" not to be opened automatically.